Fix zalgo's minimum and maximum usage
The combining marks were being overwritten every loop, so only the last one in the loop was kept. Regardless of minimum and maximum, it only ever added one combining mark to each letter. I just moved the creation base letter to above the loop, and moved the adding of the fully combined form to the translations dictionary to below the loop.